app/requirements.txt
pandas==0.25.1
python-dotenv==0.15.0
psycopg2
requests
